requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.32.3
simple_chatwoot @ git+https://github.com/fulit103/simple-chatwoot.git@89c8125e880316142d1c8b6fe10b628f53ac0b73
httpx==0.27.0